requirements.txt
python-fasthtml
fasthtml-hf
uvicorn>=0.29
groq
python-dotenv
wikipedia
duckduckgo_search